Add assertion to ensure get command arguments are used correctly
When retrieving arguments, this function should only be used for
commands that require arguments.

To prevent misuse, an assertion has been added to ensure the
argument count is greater than zero.

assert the count must be > 0 when using this function

This is the simplest way to enforce correct usage and prevent
errors caused by improper function calls.